Create Enhanced For loop.java
Java also includes another version of for loop introduced in Java 5. Enhanced for loop provides a simpler way to iterate through the elements of a collection or an array. It is inflexible and should be used only when there is a need to iterate through the elements in a sequential manner without knowing the index of the currently processed element. Also, note that the object/variable is immutable when enhanced for loop is used i.e., it ensures that the values in the array can not be modified, so it can be said as a read-only loop where you can't update the values as opposed to other loops where values can be modified. We recommend using this form of the for statement instead of the general form whenever possible.(as per JAVA doc.) Syntax:
for (T element:Collection obj/array)
{
    statement(s)
}

Let's take an example to demonstrate how enhanced for loop can be used to simplify the work. Suppose there is an array of names and we want to print all the names in that array. Let's see the difference between these two examples. Enhanced for loop simplifies the work as follows-