Create Check Palindrome.java
Given a string s, you need to check if it is palindrome or not. A palidrome is a string that reads the same from front and back.
Ignore the case in this question.

Example 1:

Input:
s = "Hello"
Output: 
false
Explanation: Hello is not equal to olleH
so it's not a palindrome.
Example 2:

Input:
s = "TenEt"
Output:
true
Explanation: TenEt = tEneT as we are
ignoring the case.