Create Count Odd Even in Array - .Java
Given an array A of N elements. The task is to count a number of even and odd elements in the array.

Example 1:

Input:
5
1 2 3 4 5

Output:
3 2

Explanation:
In the given array, there are 3 odd elements 
(1, 3, 5) and 2 even elements (2 and 4).

Example 2:
Input:
4
1 6 18 7

Output:
2 2

Explanation:
In the given array, there are 2 odd
elements (1, 7) and 2 even elements (6 and 18).