Create find Nice Array.java
you are given an array arr of n integer. an array is nice if for each index I the value of arr[I]= i. your task is to find the length of the longest subarray. which is nice. consider 1-based indexing : 
Input:
n = 7
arr = {2,1,2,3,1,2,2}
Output:
3
Explanation:
The subarray {1,2,3} is the longest nice subarray.


Input:
n = 4
arr = {3,5,2,6}
Output:
0
Explanation:
There is no nice subarray.